Move role files to separate folder.
